app/requirements.txt
Django==3.2
django-rest-framework==0.1.0  